Revert "feat: create and use madara-mapping-sync worker"
This reverts commit 53b0044de6d80fcc0bad7bc83e958ac88e3f73e5.
