--- crates/client/mapping-sync/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "madara-mapping-sync"
-version = "0.1.0"
-description = "Mapping sync logic for Madara"
-authors = [
-	"Timothée Delabrouille <https://github.com/tdelabro>",
-	"Substrate DevHub <https://github.com/substrate-developer-hub>",
-]
-homepage = "https://github.com/keep-starknet-strange/madara"
-edition = "2021"
-license = "MIT"
-publish = false
-repository = "https://github.com/keep-starknet-strange/madara"
-
-[dependencies]
-futures = "0.3.25"
-futures-timer = "3.0.1"
-log = "0.4.17"
-# Substrate
-sc-client-api = { workspace = true }
-sp-api = { workspace = true }
-sp-blockchain = { workspace = true }
-sp-runtime = { workspace = true }
-madara-db = { path = "../db" }
-pallet-starknet = { path = "../../pallets/starknet" }
